refactor crdwatcher
