+++ fiaas_deploy_daemon/__init__.py
@@ -204,7 +204,7 @@ def main():
     try:
         log.info("fiaas-deploy-daemon starting with configuration {!r}".format(cfg))
         if cfg.enable_crd_support:
-            crd_binding = CustomResourceDefinitionBindings(cfg.use_apiextensionsv1_crd)
+            crd_binding = CustomResourceDefinitionBindings()
         else:
             crd_binding = DisabledCustomResourceDefinitionBindings()
         binding_specs = [
