+++ fiaas_deploy_daemon/crd/apiextensionsv1_crd_bootstrap.py
@@ -1,0 +1,106 @@
+# coding: utf-8
+
+# Copyright 2017-2019 The FIAAS Authors
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#      http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+from __future__ import absolute_import
+
+import logging
+
+from k8s.models.common import ObjectMeta
+from k8s.models.apiextensions_v1_custom_resource_definition import CustomResourceConversion,\
+    CustomResourceDefinitionNames, CustomResourceDefinitionSpec, CustomResourceDefinition,\
+    CustomResourceDefinitionVersion, CustomResourceValidation, JSONSchemaProps
+
+LOG = logging.getLogger(__name__)
+
+
+class ApiextensionsV1CrdBootstrapper(object):
+    def __init__(self):
+        def _create(kind, plural, short_names, group, schema_properties):
+            name = "%s.%s" % (plural, group)
+            metadata = ObjectMeta(name=name)
+            names = CustomResourceDefinitionNames(kind=kind, plural=plural, shortNames=short_names)
+            schema = CustomResourceValidation(openAPIV3Schema=JSONSchemaProps(type="object", properties=schema_properties))
+            version_v1 = CustomResourceDefinitionVersion(name="v1", served=True, storage=True, schema=schema)
+            spec = CustomResourceDefinitionSpec(
+                group=group,
+                names=names,
+                versions=[version_v1],
+                preserveUnknownFields=False,
+                scope="Namespaced",
+                conversion=CustomResourceConversion(strategy="None")
+            )
+            definition = CustomResourceDefinition.get_or_create(metadata=metadata, spec=spec)
+            definition.save()
+            LOG.info("Created CustomResourceDefinition with name %s", name)
+
+        def bootstrap():
+            object_with_unknown_fields = {"type": "object", "x-kubernetes-preserve-unknown-fields": True}
+            application_schema_properties = {
+                "spec": {
+                    "type": "object",
+                    "properties": {
+                        "application": {
+                            "type": "string",
+                        },
+                        "image": {
+                            "type": "string",
+                        },
+                        "config": object_with_unknown_fields,
+                        "additional_labels": {
+                            "type": "object",
+                            "properties": {
+                                "global": object_with_unknown_fields,
+                                "deployment": object_with_unknown_fields,
+                                "horizontal_pod_autoscaler": object_with_unknown_fields,
+                                "ingress": object_with_unknown_fields,
+                                "service": object_with_unknown_fields,
+                                "service_account": object_with_unknown_fields,
+                                "pod": object_with_unknown_fields,
+                                "status": object_with_unknown_fields,
+                            }
+                        },
+                        "additional_annotations": {
+                            "type": "object",
+                            "properties": {
+                                "global": object_with_unknown_fields,
+                                "deployment": object_with_unknown_fields,
+                                "horizontal_pod_autoscaler": object_with_unknown_fields,
+                                "ingress": object_with_unknown_fields,
+                                "service": object_with_unknown_fields,
+                                "service_account": object_with_unknown_fields,
+                                "pod": object_with_unknown_fields,
+                                "status": object_with_unknown_fields,
+                            }
+                        }
+                    }
+                }
+            }
+            application_status_schema_properties = {
+                "result": {
+                    "type": "string"
+                },
+                "logs": {
+                    "type": "array",
+                    "items": {
+                        "type": "string"
+                    }
+                }
+            }
+            _create("Application", "applications", ("app", "fa"), "fiaas.schibsted.io", application_schema_properties)
+            _create("ApplicationStatus", "application-statuses", ("status", "appstatus", "fs"), "fiaas.schibsted.io",
+                        application_status_schema_properties)
+
+        bootstrap()
