+++ fiaas_deploy_daemon/crd/watcher.py
@@ -21,17 +21,16 @@ import logging
 from k8s.base import WatchEvent
 from k8s.client import NotFound
 from k8s.models.common import ObjectMeta
-from k8s.models.custom_resource_definition import CustomResourceDefinition, CustomResourceDefinitionSpec, \
-    CustomResourceDefinitionNames
 from k8s.watcher import Watcher
 from yaml import YAMLError
 
 from .status import create_name
 from .types import FiaasApplication, FiaasApplicationStatus
+from .apiextensionsv1_crd_bootstrap import ApiextensionsV1CrdBootstrapper
+from .apiextensionsv1beta1_crd_bootstrap import ApiextensionsV1Beta1CrdBootstrapper
 from ..base_thread import DaemonThread
 from ..deployer import DeployerEvent
 from ..log_extras import set_extras
-from ..retry import retry_on_upsert_conflict
 from ..specs.factory import InvalidConfiguration
 
 LOG = logging.getLogger(__name__)
@@ -46,38 +45,30 @@ class CrdWatcher(DaemonThread):
         self._lifecycle = lifecycle
         self.namespace = config.namespace
         self.enable_deprecated_multi_namespace_support = config.enable_deprecated_multi_namespace_support
+        if config.use_apiextensionsv1_crd:
+            self.crd_bootstrapper = ApiextensionsV1CrdBootstrapper
+        else:
+            self.crd_bootstrapper = ApiextensionsV1Beta1CrdBootstrapper
 
     def __call__(self):
+        self._bootstrap_custom_resource_definitions()
         while True:
             if self.enable_deprecated_multi_namespace_support:
                 self._watch(namespace=None)
             else:
                 self._watch(namespace=self.namespace)
 
+    def _bootstrap_custom_resource_definitions(self):
+        self.crd_bootstrapper()
+
     def _watch(self, namespace):
         try:
             for event in self._watcher.watch(namespace=namespace):
                 self._handle_watch_event(event)
         except NotFound:
-            self.create_custom_resource_definitions()
+            self._bootstrap_custom_resource_definitions()
         except Exception:
             LOG.exception("Error while watching for changes on FiaasApplications")
-
-    @classmethod
-    def create_custom_resource_definitions(cls):
-        cls._create("Application", "applications", ("app", "fa"), "fiaas.schibsted.io")
-        cls._create("ApplicationStatus", "application-statuses", ("status", "appstatus", "fs"), "fiaas.schibsted.io")
-
-    @staticmethod
-    @retry_on_upsert_conflict
-    def _create(kind, plural, short_names, group):
-        name = "%s.%s" % (plural, group)
-        metadata = ObjectMeta(name=name)
-        names = CustomResourceDefinitionNames(kind=kind, plural=plural, shortNames=short_names)
-        spec = CustomResourceDefinitionSpec(group=group, names=names, version="v1")
-        definition = CustomResourceDefinition.get_or_create(metadata=metadata, spec=spec)
-        definition.save()
-        LOG.info("Created CustomResourceDefinition with name %s", name)
 
     def _handle_watch_event(self, event):
         if event.type in (WatchEvent.ADDED, WatchEvent.MODIFIED):
