+++ tests/fiaas_deploy_daemon/test_bootstrap_e2e.py
@@ -28,7 +28,6 @@ from k8s.models.deployment import Deployment
 from k8s.models.ingress import Ingress
 from k8s.models.service import Service
 
-from fiaas_deploy_daemon.crd.apiextensionsv1_crd_watcher import ApiextensionsV1CrdWatcher
 from fiaas_deploy_daemon.crd.types import FiaasApplication, FiaasApplicationSpec
 from fiaas_deploy_daemon.crd.watcher import CrdWatcher
 from fiaas_deploy_daemon.tools import merge_dicts
@@ -139,12 +138,6 @@ class TestBootstrapE2E(object):
 
     def test_bootstrap_crd(self, request, kubernetes, k8s_version, use_docker_for_e2e):
         skip_if_crd_not_supported(k8s_version)
-
-        if apiextensions_v1beta1_crd_supported(k8s_version):
-            CrdWatcher.create_custom_resource_definitions()
-        elif apiextensions_v1_crd_supported(k8s_version):
-            ApiextensionsV1CrdWatcher.create_custom_resource_definitions()
-        wait_until(crd_available(kubernetes, timeout=TIMEOUT), "CRD available", RuntimeError, patience=PATIENCE)
 
         def prepare_test_case(test_case):
             name, fiaas_application, expected = self.custom_resource_definition_test_case(*test_case)
